+++ Sources/SwiftInfoCore/Output.swift
@@ -3,10 +3,12 @@ import Foundation
 public struct Output {
     let rawDictionary: [String: Any]
     let summaries: [Summary]
+    let errors: [String]
 
     init<T: InfoProvider>(info: ExtractedInfo<T>) throws {
         self.rawDictionary = try info.encoded()
         self.summaries = [info.summary]
+        self.errors = []
     }
 
     func extractedInfo<T: InfoProvider>(ofType type: T.Type) throws -> T? {
@@ -26,15 +28,18 @@ extension Output {
         let dict = lhsDict.merging(rhsDict) { new, _ in
             return new
         }
-        return Output(rawDictionary: dict, summaries: lhs.summaries + rhs.summaries)
+        return Output(rawDictionary: dict,
+                      summaries: lhs.summaries + rhs.summaries,
+                      errors: lhs.errors + rhs.errors)
     }
 
     public static func +=(lhs: inout Output, rhs: Output) {
         lhs = lhs + rhs
     }
 
-    public init(rawDictionary: [String: Any], summaries: [Summary]) {
+    public init(rawDictionary: [String: Any], summaries: [Summary], errors: [String]) {
         self.rawDictionary = rawDictionary
         self.summaries = summaries
+        self.errors = errors
     }
 }