+++ Sources/SwiftInfoCore/Runner.swift
@@ -16,7 +16,7 @@ public enum Runner {
             "-fmodule-map-file=\(fileUtils.toolFolder)Csourcekitd/include/module.modulemap",
             "-I",
             "\(fileUtils.toolFolder)Csourcekitd/include",
-            fileUtils.infofileFolder + "Infofile.swift",
+            (try! fileUtils.infofileFolder()) + "Infofile.swift",
             "-toolchain",
             "\(toolchainPath)"] + Array(processInfoArgs.dropFirst()) // Route SwiftInfo args to the sub process
     }